Push dev branch to main

+++ src/services/email_verification.py
@@ -30,7 +30,7 @@ class EmailVerificationService:
        )
 
     try:
-       user = get_user(session=session, email=payload.email.lower())
+      user = get_user(session=session, email=payload.email.lower())
     except Exception:
        return False
     
@@ -61,8 +61,11 @@ class EmailVerificationService:
    def purge_user_otp(self, session: Session, email: str):
       # Delete all user's OTP from otps table
       deleted_rows = session.query(OTP).filter(OTP.email == email)
-      deleted_rows.delete()
-      session.commit()
+
+      # Purge all OTP related to the user, if any
+      if deleted_rows:
+         deleted_rows.delete()
+         session.commit()
       
       return deleted_rows or None
 
@@ -83,8 +86,8 @@ class EmailVerificationService:
          print(f"Error while inserting token to DB: {e}")
          session.rollback()
          raise e
-      finally:
-         session.close()
+      # finally:
+      #    session.close()
 
    def get_latest_valid_otp(self, session: Session, email: str):
       """
